refactor(frontend): découpe MergePreviewModal et useComicForm en sous-composants et hooks ciblés
- MergePreviewModal (587→80 lignes) : MergeMetadataForm + MergeTomeTable + useMergePreviewForm (useReducer)
- useComicForm (420→210 lignes) : useLookupFeature + useTomeManagement + useAuthorManagement
- TomeTable : objet TomeManager au lieu de 12 props
- SyncFailureSection extrait dans components/

fixes #268
